openexr 3.2.1
This dragged a new dependency libdeflate 1.18 which it was determined
to either download itself from git which is a hard no for us, or use
pkg_config to find it, which windows doesn't use/have.

we were not the first ones to run into this, someone posted a
patch in an upstream ticket over at

https://github.com/AcademySoftwareFoundation/openexr/issues/1588

patch was used almost used verbatim with the sole change being
using the static version for the following reasons:

- there are currently no other consumers of this library
- I'm lazy and short on time, it being static doesn't require any
  work harvesting and distributing the build artefacts on the blender
  side of things.

I picked deflate 1.18 since it's what openexr defaults 1.19 is
available though, unsure what we want to do here.